Более менее всю логику сделал, теперь рефакторинг: начал с UserService метод logOut  - на вход уже должны приходить всегда валидные данные, внутри только проверка на основе этих данных на наличие пользователя и тд. Парисинга здесь нет, эти сервисы могут в дальнейшем работать с другим сервером, логика их работы неизменна..парсинг джсон строки происходит в сервере, формат в дальнейшем может быть другой, проверка на валидность формата здесь же

+++ src/main/java/SongService.java
@@ -6,7 +6,16 @@ import model.Rating;
 import model.Song;
 import model.User;
 import request.DeleteSongDtoRequest;
+import request.FindByComposersDtoRequest;
+import request.GetSongsDtoRequest;
 import request.RegisterSongDtoRequest;
+import response.ErrorDtoResponse;
+import response.FindByComposersDtoResponse;
+import response.GetSongsDtoResponse;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 
 public class SongService {
 
@@ -24,11 +33,11 @@ public class SongService {
                            String token) {
         Song song = new Song();
         song.setSongName(songName);
-        song.setComposer(composer);
+        song.setComposers(composer);
         song.setAuthor(author);
         song.setMusician(musician);
         song.setSongDuration(songDuration);
-        song.setLogin(db.getUserByToken(token).getLogin());
+        song.setLogin(userService.getUserByToken(token).getLogin());
         song.setSongId(1);
         return song;
     }
@@ -43,7 +52,8 @@ public class SongService {
      */
     public String addSong(String requestJsonString) throws Exception {
         RegisterSongDtoRequest request = mapper.readValue(requestJsonString, RegisterSongDtoRequest.class);
-        User user = db.getUserByToken(request.getToken());
+        ErrorDtoResponse response = new ErrorDtoResponse("User not found");
+        User user = userService.getUserByToken(request.getToken());
         if (user != null) {
             Song newSong = createSong(request.getSongName(), request.getComposer(), request.getAuthor(),
                     request.getMusician(), request.getSongDuration(), request.getToken());
@@ -51,7 +61,7 @@ public class SongService {
             songDao.insert(newSong);
             ratingDao.updateRating(new Rating(user.getLogin(), newSong.getSongId(), 5));
             return "{}";
-        } else return "{\"error\" : \"User not found\"}";
+        } else return mapper.writeValueAsString(response);
     }
 
     /**
@@ -61,46 +71,73 @@ public class SongService {
      * @return
      */
     public String deleteSong(String requestJsonString) throws Exception {
-        DeleteSongDtoRequest deleteSongDtoRequest = mapper.readValue(requestJsonString, DeleteSongDtoRequest.class);
-        System.out.println(db.frequencyRaitings(deleteSongDtoRequest.getSongId()));
-        Song song = db.findSongById(deleteSongDtoRequest.getSongId());
+        DeleteSongDtoRequest request = mapper.readValue(requestJsonString, DeleteSongDtoRequest.class);
+        Song song = db.findSongById(request.getSongId());
+        User user = userService.getUserByToken(request.getToken());
+        if (user == null) {
+            ErrorDtoResponse response = new ErrorDtoResponse("User not found");
+            return mapper.writeValueAsString(response);
+        }
         if (song == null) {
-            return "{\"error\" : \"Song not found\"}";
+            ErrorDtoResponse response = new ErrorDtoResponse("Song not found");
+            return mapper.writeValueAsString(response);
         }
         return songDao.deleteSong(song);
-
     }
 
-    //прежде чем добавить оценку песне, нужно её сначала найти
-    public Song findSongByAuthor(String author) {
-        DataBase db = DataBase.getInstance();
-        for (Song song : db.getSongList()) {
-            if (song.getAuthor().equals(author)) {
-                return song;
+    public String findSongByComposer(String requestJsonString) throws Exception {
+        List<Song> songs = new ArrayList<Song>();
+        FindByComposersDtoRequest request = mapper.readValue(requestJsonString, FindByComposersDtoRequest.class);
+        User user = userService.getUserByToken(request.getToken());
+        if (user == null) {
+            return mapper.writeValueAsString(new ErrorDtoResponse("User not found"));
+        }
+        for (Song song : songDao.getSongList()) {
+            if (Arrays.asList(song.getComposers()).containsAll(Arrays.asList(request.getComposers()))) {
+                songs.add(song);
             }
         }
-        return null;
+//        for (int i = 0; i < songDao.getSongList().get(i).getComposers().length; i++) {
+//            for (int j = 0; j < request.getComposers().length; j++) {
+//                if (songDao.getSongList().get(i).getComposers()[i].equals(request.getComposers()[j])) {
+//                    songs.add(songDao.getSongList().get(i));
+//                }
+//            }
+//        }
+        FindByComposersDtoResponse response = new FindByComposersDtoResponse();
+        response.getSongs().addAll(songs);
+        if (songs.size() == 0) {
+            ErrorDtoResponse errorResponse = new ErrorDtoResponse("Composers not found");
+            return mapper.writeValueAsString(errorResponse);
+        }
+        return mapper.writerWithDefaultPrettyPrinter().writeValueAsString(response);
     }
 
     private void generateSongId(Song song) {
-        //берется последняя песня и к её songId добавляется единица
         if (db.getSongList().size() != 0) {
             int newId = db.getSongList().get(db.getSongList().size() - 1).getSongId() + 1;
             song.setSongId(newId);
-        }
-        else return;
+        } else return;
     }
 
     /**
      * Радиослушатель  получает список песен. requestJsonString содержит параметры для отбора песен и token,
      * полученный как результат выполнения команды регистрации радиослушателя. Метод при успешном выполнении возвращает
      * json с описанием всех песен. Если же команду почему-то выполнить нельзя, возвращает json с элементом “error”
-     * @param requestJsonString
-     * @return
      *
-     * Как узнать, что в параметрах для отбора песен именно автор песни, а не композитор? где его искать?
+     * @param requestJsonString
+     * @return Как узнать, что в параметрах для отбора песен именно автор песни, а не композитор? где его искать?
      */
 
-    public String getSongs(String requestJsonString) {
-        return "fig";
-    }}
+    public String getSongs(String requestJsonString) throws Exception {
+        GetSongsDtoRequest getSongsDtoRequest = mapper.readValue(requestJsonString, GetSongsDtoRequest.class);
+        User user = userService.getUserByToken(getSongsDtoRequest.getToken());
+        if (user != null) {
+            GetSongsDtoResponse getSongsDtoResponse = new GetSongsDtoResponse();
+            getSongsDtoResponse.getSongList().addAll(songDao.getSongList());
+            return mapper.writerWithDefaultPrettyPrinter().writeValueAsString(getSongsDtoResponse);
+        }
+        ErrorDtoResponse response = new ErrorDtoResponse("User not found");
+        return mapper.writeValueAsString(response);
+    }
+}
